Move text from spec to actual enigma file.
